layout/generic/test/test_bug404872.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=404872
-->
<head>
  <title>Test for Bug 404872</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=404872">Mozilla Bug 404872</a>
<p id="display">
  <select multiple id="a">
    <option name="value1">value1</option>
    <option name="value2">value2</option>
    <option name="value3">value3</option>
  </select>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/* Focus, press key down twice, and make sure we successfully selected the second item. */
$('a').focus();
synthesizeKey("KEY_ArrowDown");
synthesizeKey("KEY_ArrowDown");
is($("a").value, "value2", "value2 should be selected!");


</script>
</pre>
</body>
</html>
